Do not delete the checkpoint model

+++ flowdiffusion/train_policy/train_policy_calvin.py
@@ -377,7 +377,8 @@ def main(args):
                         results_folder, f"model-{step // cfg.save_every - 2}.pt"
                     )
                     if os.path.exists(past_saving_path):
-                        os.remove(past_saving_path)
+                        if step // cfg.save_every - 2 != args.checkpoint_num:
+                            os.remove(past_saving_path)
                     # Save model
                     saving_path = os.path.join(
                         results_folder, f"model-{step // cfg.save_every}.pt"
